+++ include/operations/ImageOperation.h
