+++ include/operations/channel-operations/FastFourierOperation.h
@@ -1,17 +1,18 @@
 //
 // Created by gluckasz on 2/4/25.
 //
-#include "../../include/ImageOperation.h"
-#include "../../include/input-processing-lib/OutputManager.h"
+
+#include "../ImageOperation.h"
+#include "input-processing-lib/OutputManager.h"
 #include "image-processing-lib/FourierProcessor.h"
 
-class FourierOperation final : public ImageOperation {
+class FastFourierOperation final : public ImageOperation {
 public:
     void apply(cv::Mat &image) const override {
-        cv::Mat fourierImage = FourierProcessor::fourierTransform(image);
+        cv::Mat fourierImage = FourierProcessor::fastFourierTransform(image);
         const std::string path = OutputManager::constructPath("image_fourier", "magnitude_spectrum", "bmp");
         FourierProcessor::visualizeFourier(fourierImage, path);
-        image = FourierProcessor::inverseFourierTransform(fourierImage);
+        image = FourierProcessor::inverseFastFourierTransform(fourierImage);
     }
 };
