+++ include/operations/whole-image-operations/ErosionOperation.h
